Fixed some things that Google's cpplint warns about (mostly, constructors that should be marked explicit).

+++ source/StarField.cpp
@@ -96,16 +96,16 @@ void StarField::Draw(const Point &pos, const Point &vel) const
 	double borderX = fabs(vel.X()) + 1.;
 	double borderY = fabs(vel.Y()) + 1.;
 	// Find the absolute bounds of the star field we must draw.
-	long minX = Screen::Left() + pos.X() - borderX;
-	long minY = Screen::Top() + pos.Y() - borderY;
-	long maxX = Screen::Right() + pos.X() + borderX;
-	long maxY = Screen::Bottom() + pos.Y() + borderY;
+	int minX = Screen::Left() + pos.X() - borderX;
+	int minY = Screen::Top() + pos.Y() - borderY;
+	int maxX = Screen::Right() + pos.X() + borderX;
+	int maxY = Screen::Bottom() + pos.Y() + borderY;
 	// Round down to the start of the nearest tile.
 	minX &= ~(TILE_SIZE - 1l);
 	minY &= ~(TILE_SIZE - 1l);
 	
-	for(long gy = minY; gy < maxY; gy += TILE_SIZE)
-		for(long gx = minX; gx < maxX; gx += TILE_SIZE)
+	for(int gy = minY; gy < maxY; gy += TILE_SIZE)
+		for(int gx = minX; gx < maxX; gx += TILE_SIZE)
 		{
 			Point off = Point(gx, gy) - pos;
 			GLfloat translate[2] = {
@@ -219,15 +219,15 @@ void StarField::MakeStars(int stars, int width)
 	tileIndex.clear();
 	tileIndex.resize(tileCols * tileCols, 0);
 	
-	vector<short> off;
-	static const short MAX_OFF = 50;
-	static const short MAX_D = MAX_OFF * MAX_OFF;
-	static const short MIN_D = MAX_D / 4;
+	vector<int> off;
+	static const int MAX_OFF = 50;
+	static const int MAX_D = MAX_OFF * MAX_OFF;
+	static const int MIN_D = MAX_D / 4;
 	off.reserve(MAX_OFF * MAX_OFF * 5);
-	for(short x = -MAX_OFF; x <= MAX_OFF; ++x)
-		for(short y = -MAX_OFF; y <= MAX_OFF; ++y)
+	for(int x = -MAX_OFF; x <= MAX_OFF; ++x)
+		for(int y = -MAX_OFF; y <= MAX_OFF; ++y)
 		{
-			short d = x * x + y * y;
+			int d = x * x + y * y;
 			if(d < MIN_D || d > MAX_D)
 				continue;
 			
@@ -237,11 +237,11 @@ void StarField::MakeStars(int stars, int width)
 	
 	// Generate random points in a temporary vector.
 	// Keep track of how many fall into each tile, for sorting out later.
-	vector<short> temp;
+	vector<int> temp;
 	temp.reserve(2 * stars);
 	
-	short x = Random::Int(width);
-	short y = Random::Int(width);
+	int x = Random::Int(width);
+	int y = Random::Int(width);
 	for(int i = 0; i < stars; ++i)
 	{
 		for(int j = 0; j < 10; ++j)
@@ -269,12 +269,12 @@ void StarField::MakeStars(int stars, int width)
 	for(auto it = temp.begin(); it != temp.end(); )
 	{
 		// Figure out what tile this star is in.
-		short x = *it++;
-		short y = *it++;
+		int x = *it++;
+		int y = *it++;
 		int index = (x / TILE_SIZE) + (y / TILE_SIZE) * tileCols;
 		
 		// Randomize its sub-pixel position and its size / brightness.
-		short random = Random::Int(4096);
+		int random = Random::Int(4096);
 		float fx = (x & (TILE_SIZE - 1)) + (random & 15) * 0.0625f;
 		float fy = (y & (TILE_SIZE - 1)) + (random >> 8) * 0.0625f;
 		float size = (((random >> 4) & 15) + 20) * 0.0625f;
